+++ app/src/main/java/archdesign/response/JsonOutputSchema.java
@@ -50,22 +50,22 @@ public class JsonOutputSchema {
      */
     public static class OversizedPiece {
         @SerializedName("side1")
-        private int side1;
+        private double side1;
         
         @SerializedName("side2")
-        private int side2;
+        private double side2;
         
         @SerializedName("quantity")
         private int quantity;
         
-        public OversizedPiece(int side1, int side2, int quantity) {
+        public OversizedPiece(double side1, double side2, int quantity) {
             this.side1 = side1;
             this.side2 = side2;
             this.quantity = quantity;
         }
         
-        public int getSide1() { return side1; }
-        public int getSide2() { return side2; }
+        public double getSide1() { return side1; }
+        public double getSide2() { return side2; }
         public int getQuantity() { return quantity; }
     }
